+++ src/main/java/com/capston/favicon/user/application/AuthServiceImpl.java
@@ -1,8 +1,8 @@
-package com.capston.favicon.application;
+package com.capston.favicon.user.application;
 
-import com.capston.favicon.application.repository.AuthService;
-import com.capston.favicon.domain.domain.User;
-import com.capston.favicon.infrastructure.UserRepository;
+import com.capston.favicon.user.application.service.AuthService;
+import com.capston.favicon.user.domain.User;
+import com.capston.favicon.user.infrastructure.UserRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
