+++ src/main/java/com/capston/favicon/user/application/service/AuthService.java
@@ -1,4 +1,4 @@
-package com.capston.favicon.application.repository;
+package com.capston.favicon.user.application.service;
 
 import jakarta.servlet.http.HttpServletRequest;
